Extract ServerFallback component
